Solve challenge 6
This could be improved upon by changing the key length detection
to be more robust, by analysing more multiples of the text such that
the "real keysize" gets the lowest score.
The code for this isn't super complex, just basically removing that
DRY section of code with a dynamic version.